[Feat] Random Road, House
도로, 집 배치(수정 필요)

+++ Assets/3.Script/Donggil/Map/HouseType.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class HouseType
 {
     [SerializeField] private GameObject[] prefabs;
-    public int sizeRequired;
+    public float sizeRequired;
     public int quantity;
     public int quantityAlreadyPlaced;
